Fix ujint to be unsigned long
This removes a warning involving a printf().

1. Also, rename app_AlphaTab() to cat_cstr_AlphaTab().
  c. alphatab.h  cat_cstr_AlphaTab()
  c. cx.c

+++ def.h
@@ -24,7 +24,7 @@ typedef unsigned char byte;
 typedef unsigned int uint;
 #define Max_uint UINT_MAX
 
-typedef size_t ujint;
+typedef unsigned long int ujint;
 #define Max_ujint ((ujint)0 - 1)
 
 #if 0
